DDF-4901 more support for react
