+++ ui/packages/catalog-ui-search/src/main/webapp/component/query-editor/query-editor.view.js
@@ -13,116 +13,16 @@
  *
  **/
 
+import * as React from 'react'
 const Marionette = require('marionette')
 const template = require('./query-editor.hbs')
 const CustomElements = require('../../js/CustomElements.js')
-const QueryBasic = require('../query-basic/query-basic.view.js')
 const QueryAdvanced = require('../query-advanced/query-advanced.view.js')
 const QueryTitle = require('../query-title/query-title.view.js')
-const QueryAdhoc = require('../query-adhoc/query-adhoc.view.js')
-const cql = require('../../js/cql.js')
-const CQLUtils = require('../../js/CQLUtils.js')
 const store = require('../../js/store.js')
 const announcement = require('../announcement/index.jsx')
 import { InvalidSearchFormMessage } from 'component/announcement/CommonMessages'
-
-function isNested(filter) {
-  let nested = false
-  filter.filters.forEach(subfilter => {
-    nested = nested || subfilter.filters
-  })
-  return nested
-}
-
-function isTypeLimiter(filter) {
-  let typesFound = {}
-  filter.filters.forEach(subfilter => {
-    typesFound[CQLUtils.getProperty(subfilter)] = true
-  })
-  typesFound = Object.keys(typesFound)
-  return (
-    typesFound.length === 2 &&
-    typesFound.indexOf('metadata-content-type') >= 0 &&
-    typesFound.indexOf('datatype') >= 0
-  )
-}
-
-function isAnyDate(filter) {
-  const propertiesToCheck = [
-    'created',
-    'modified',
-    'effective',
-    'metacard.created',
-    'metacard.modified',
-  ]
-  const typesFound = {}
-  const valuesFound = {}
-  if (filter.filters.length === propertiesToCheck.length) {
-    filter.filters.forEach(subfilter => {
-      typesFound[subfilter.type] = true
-      valuesFound[subfilter.value] = true
-      const indexOfType = propertiesToCheck.indexOf(
-        CQLUtils.getProperty(subfilter)
-      )
-      if (indexOfType >= 0) {
-        propertiesToCheck.splice(indexOfType, 1)
-      }
-    })
-    return (
-      propertiesToCheck.length === 0 &&
-      Object.keys(typesFound).length === 1 &&
-      Object.keys(valuesFound).length === 1
-    )
-  }
-  return false
-}
-
-function translateFilterToBasicMap(filter) {
-  const propertyValueMap = {}
-  let downConversion = false
-  if (filter.filters) {
-    filter.filters.forEach(filter => {
-      if (!filter.filters) {
-        propertyValueMap[CQLUtils.getProperty(filter)] =
-          propertyValueMap[CQLUtils.getProperty(filter)] || []
-        if (
-          propertyValueMap[CQLUtils.getProperty(filter)].filter(
-            existingFilter => existingFilter.type === filter.type
-          ).length === 0
-        ) {
-          propertyValueMap[CQLUtils.getProperty(filter)].push(filter)
-        }
-      } else if (!isNested(filter) && isAnyDate(filter)) {
-        propertyValueMap['anyDate'] = propertyValueMap['anyDate'] || []
-        if (
-          propertyValueMap['anyDate'].filter(
-            existingFilter => existingFilter.type === filter.filters[0].type
-          ).length === 0
-        ) {
-          propertyValueMap['anyDate'].push(filter.filters[0])
-        }
-      } else if (!isNested(filter) && isTypeLimiter(filter)) {
-        propertyValueMap[CQLUtils.getProperty(filter.filters[0])] =
-          propertyValueMap[CQLUtils.getProperty(filter.filters[0])] || []
-        filter.filters.forEach(subfilter => {
-          propertyValueMap[CQLUtils.getProperty(filter.filters[0])].push(
-            subfilter
-          )
-        })
-      } else {
-        downConversion = true
-      }
-    })
-  } else {
-    propertyValueMap[CQLUtils.getProperty(filter)] =
-      propertyValueMap[CQLUtils.getProperty(filter)] || []
-    propertyValueMap[CQLUtils.getProperty(filter)].push(filter)
-  }
-  return {
-    propertyValueMap,
-    downConversion,
-  }
-}
+import ExtensionPoints from '../../extension-points'
 
 module.exports = Marionette.LayoutView.extend({
   template,
@@ -154,19 +54,19 @@ module.exports = Marionette.LayoutView.extend({
     }
   },
   reshow() {
-    switch (this.model.get('type')) {
-      case 'text':
-        this.showText()
-        break
-      case 'basic':
-        this.showBasic()
-        break
-      case 'advanced':
-        this.showAdvanced()
-        break
+    this.queryView = undefined
+    const formType = this.model.get('type')
+    switch (formType) {
       case 'custom':
         this.showCustom()
         break
+      default:
+        const queryForm = ExtensionPoints.queryForms.find(
+          form => form.id === formType
+        )
+        if (queryForm) {
+          this.showQueryForm(queryForm)
+        }
     }
     this.edit()
   },
@@ -181,25 +81,24 @@ module.exports = Marionette.LayoutView.extend({
       })
     )
   },
-  showText() {
-    const translationToBasicMap = translateFilterToBasicMap(
-      cql.simplify(cql.read(this.model.get('cql')))
-    )
-    this.queryContent.show(
-      new QueryAdhoc({
-        model: this.model,
-        text: translationToBasicMap.propertyValueMap.anyText
-          ? translationToBasicMap.propertyValueMap.anyText[0].value
-          : '',
-      })
-    )
-  },
-  showBasic() {
-    this.queryContent.show(
-      new QueryBasic({
-        model: this.model,
-      })
-    )
+  showQueryForm(form) {
+    const options = form.options || {}
+    const queryFormView = form.reactView
+      ? new (Marionette.LayoutView.extend({
+          template: () => (
+            <form.reactView
+              model={this.model}
+              options={options}
+              onRef={ref => (this.queryView = ref)}
+            />
+          ),
+        }))({})
+      : new form.view({
+          model: this.model,
+          ...options,
+        })
+
+    this.queryContent.show(queryFormView)
   },
   showCustom() {
     this.queryContent.show(
@@ -224,18 +123,23 @@ module.exports = Marionette.LayoutView.extend({
   },
   edit() {
     this.$el.addClass('is-editing')
-    this.queryContent.currentView.edit()
+    this.queryView
+      ? this.queryView.edit()
+      : this.queryContent.currentView.edit()
   },
   cancel() {
     this.$el.removeClass('is-editing')
     this.onBeforeShow()
   },
   save() {
-    if (!this.queryContent.currentView.isValid()) {
+    const queryContentView = this.queryView
+      ? this.queryView
+      : this.queryContent.currentView
+    if (!queryContentView.isValid()) {
       announcement.announce(InvalidSearchFormMessage)
       return
     }
-    this.queryContent.currentView.save()
+    queryContentView.save()
     this.queryTitle.currentView.save()
     if (store.getCurrentQueries().get(this.model) === undefined) {
       store.getCurrentQueries().add(this.model)
@@ -245,11 +149,14 @@ module.exports = Marionette.LayoutView.extend({
     this.originalType = this.model.get('type')
   },
   saveRun() {
-    if (!this.queryContent.currentView.isValid()) {
+    const queryContentView = this.queryView
+      ? this.queryView
+      : this.queryContent.currentView
+    if (!queryContentView.isValid()) {
       announcement.announce(InvalidSearchFormMessage)
       return
     }
-    this.queryContent.currentView.save()
+    queryContentView.save()
     this.queryTitle.currentView.save()
     if (store.getCurrentQueries().get(this.model) === undefined) {
       store.getCurrentQueries().add(this.model)
